Split mesos auth into standalone class
Split mesos auth into a standalone class so you can get an
authentication token without doing any of the other mesos-related
activities.

The auth work is now in mesos_auth.cpp and has the authenticate() and
refresh_token() methods that used to be in mesos.cpp and a new method
get_token() that returns the token.

mesos.cpp had an unused bool m_token_authentication which was removed
and not carried over.